Initial Commit.
This actually is a much farther commit than initial. Much of the server
code from polka and wildcard is working correctly. The client that
consumes them though isn't yet working.

Hegel also is a bit in a broken state. The polka types are from
Typescript but, don't seem to be working quite right in the Hegel world.
The endpoints themselves also don't have any decorated types.