Serve Static Built Frontend Files.
Utilize Parcel to bundle/build front end files. Take these files and
serve them statically using the sirv package. This allows the
application to be built for a list of browsers and be served through
Polka. This eases any CORS or preflight issues that might arise from
serving the client and the server on different domains/ports.